requirements-s.txt
speedtest-cli>=2.1.3
pandas>=2.0.0
openpyxl>=3.1.0
